facebook icons, refactor

+++ src/app/components/profile/photo-list/photo-list-item/photo-list-item.component.ts
@@ -1,6 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, ElementRef, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { AuthenticationService } from 'src/app/components/authentication/authentication.service';
+import { ConfirmDialogService } from 'src/app/components/share/confirm-dialog/confirm-dialog.service';
 import { PhotoResponse } from 'src/app/components/share/photo/photo-response';
 import { PhotoService } from 'src/app/components/share/photo/photo.service';
 
@@ -14,38 +15,31 @@ export class PhotoListItemComponent implements OnInit {
   @Input() photo!: PhotoResponse;
   @Input() setId?: number;
   @Output() photoDeleted: EventEmitter<number> = new EventEmitter();
-  loggedInUserId: number;
+  principalId = this.authenticationService.getPrincipalId();
   isOptionsVisible: boolean = false;
 
   @ViewChild('optionsBtn') optionsBtn!: ElementRef;
 
   constructor(
     private authenticationService: AuthenticationService,
-    private http: HttpClient,
-    private photoService: PhotoService
+    private photoService: PhotoService,
+    private confirmService: ConfirmDialogService
   ) {
-    this.loggedInUserId = authenticationService.getPrincipalId();
   }
 
   ngOnInit(): void {
   }
 
-  showOptions(): void {
-    this.isOptionsVisible = !this.isOptionsVisible;
-  }
+  async deletePhoto() {
+    const isConfirmed = await this.confirmService.confirm({
+      title: 'Xoá hình ảnh',
+      body: 'Bạn có chắc muốn xoá ảnh này?'
+    });
 
-  hideOptions(event: any): void {
-    let btnAndChildren = [...this.optionsBtn.nativeElement.children,
-    this.optionsBtn.nativeElement];
-    if (!btnAndChildren.includes(event.target)) {
-      this.isOptionsVisible = false;
+    if (isConfirmed) {
+      this.photoService.delete(this.photo.id).subscribe(() => {
+        this.photoDeleted.emit(this.photo.id);
+      });
     }
   }
-
-  deletePhoto(): void {
-    this.photoService.delete(this.photo.id).subscribe(() => {
-      this.photoDeleted.emit(this.photo.id);
-    });
-  }
-
 }
